services/risk-analyzer-service/pyproject.toml
[project]
name = "risk-analyzer-service"
version = "0.1.0"
description = "Continuous risk analysis and adaptive rescanning prioritization service"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "uvicorn[standard]>=0.38.0",
    "pydantic-settings>=2.11.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-pubsub>=2.33.0",
    "python-json-logger>=4.0.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
